trabajo_final_bot_reciclador/requirements.txt
tensorflow==2.11.0
pyttsx3==2.90
Pillow==9.3.0
discord.py==2.3.2
python-dotenv==1.0.1
numpy==1.23.5

